feat: handle large diffs and binaries
